CLOUDP-195922: refactor E2E tests
