+++ test/e2e/atlas/compliancepolicy_setup_test.go
@@ -1,0 +1,82 @@
+// Copyright 2023 MongoDB Inc
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build e2e || (atlas && backup && compliancepolicy)
+
+package atlas_test
+
+import (
+	"encoding/json"
+	"os"
+	"os/exec"
+	"testing"
+
+	"github.com/mongodb/mongodb-atlas-cli/test/e2e"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	atlasv2 "go.mongodb.org/atlas-sdk/v20230201004/admin"
+)
+
+//nolint:thelper
+func testCompliancePolicySetup(t *testing.T, g *atlasE2ETestGenerator) {
+	cliPath, err := e2e.AtlasCLIBin()
+	r := require.New(t)
+	r.NoError(err)
+
+	scheduledPolicyItem := atlasv2.DiskBackupApiPolicyItem{
+		FrequencyInterval: 1,
+		FrequencyType:     "daily",
+		RetentionUnit:     "days",
+		RetentionValue:    1,
+	}
+
+	email := authorizedEmail
+
+	policy := &atlasv2.DataProtectionSettings{
+		ScheduledPolicyItems: []atlasv2.DiskBackupApiPolicyItem{scheduledPolicyItem},
+		ProjectId:            &g.projectID,
+		AuthorizedEmail:      &email,
+	}
+	path := "./compliancepolicy.json"
+
+	createJSONFile(t, policy, path)
+
+	t.Run("setup happy flow", func(t *testing.T) {
+		cmd := exec.Command(cliPath,
+			backupsEntity,
+			compliancepolicyEntity,
+			"setup",
+			"--projectId",
+			g.projectID,
+			"-o=json",
+			"--force",
+			"--file",
+			path,
+			"--watch",
+		)
+
+		cmd.Env = os.Environ()
+		resp, outputErr := cmd.CombinedOutput()
+
+		trimmedResponse := removeDotsFromWatching(resp)
+
+		r.NoError(outputErr, string(resp))
+		a := assert.New(t)
+
+		var result atlasv2.DataProtectionSettings
+		require.NoError(t, json.Unmarshal(trimmedResponse, &result), trimmedResponse)
+		a.Equal(1, len(result.GetScheduledPolicyItems()))
+		a.Equal(result.GetAuthorizedEmail(), authorizedEmail)
+	})
+}
